+++ tosixinch/tests/test_windows.py
@@ -7,10 +7,6 @@ import pytest
 
 from tosixinch import location
 
-@pytest.fixture(autouse=True)
-def use_ntpath(monkeypatch):
-    monkeypatch.setattr(os, 'path', ntpath)
-
 
 class TestWindowsMakePath:
 
@@ -71,33 +67,17 @@ class TestWindowsMakePath:
             self.compare(url, url, fnew)
 
 
-class TestWindowsLocalReferenceRaw:
-
-    def compare(self, url, local_url, fname):
-        comp = location.Component(url, '.', platform='win32')
-        url = location._tamper_windows_fname(url)
-        assert comp._escape_fname_reference(url) == local_url
-        assert location._tamper_fname(url) == fname
-
-    def test(self):
-        url, local_url, fname = (
-            'aaa/bbb?cc',
-            r'aaa\bbb_cc',
-            r'aaa\bbb_cc')
-        self.compare(url, local_url, fname)
-
-
 class TestWindowsLocalReference:
 
-    def compare(self, url, local_url, fname):
+    def compare(self, url, fname, ref):
         base = 'http://aaa.org'
         comp= location.Component(url, base, platform='win32')
-        assert comp.fname_reference == local_url
         assert comp.fname == fname
+        assert comp.relative_reference == ref
 
     def test(self):
-        url, local_url, fname = (
+        url, fname, ref  = (
             'https://aaa.org/bbb',
-            'bbb/_',
-            r'_htmls\aaa.org\bbb\_')
-        self.compare(url, local_url, fname)
+            r'_htmls\aaa.org\bbb\_',
+            'bbb/_')
+        self.compare(url, fname, ref)